split render from viewer move alfresco render in content pack
