+++ demo-shell/src/app/components/overlay-viewer/overlay-viewer.component.html
@@ -15,8 +15,8 @@
     (preview)="showPreview($event)">
 </adf-document-list>
 
-<!--<adf-viewer-->
-<!--    [(showViewer)]="showViewer"-->
-<!--    [overlayMode]="isOverlay"-->
-<!--    [nodeId]="nodeId">-->
-<!--</adf-viewer>-->
+<adf-alfresco-viewer
+    [showViewer]="showViewer"
+    [overlayMode]="isOverlay"
+    [nodeId]="nodeId">
+</adf-alfresco-viewer>
