+++ lib/content-services/src/lib/viewer/components/alfresco-viewer.component.scss
@@ -1,0 +1,82 @@
+/* stylelint-disable scss/at-extend-no-missing-placeholder */
+.adf-full-screen {
+    width: 100%;
+    height: 100%;
+    background-color: var(--theme-card-bg-color);
+}
+
+.adf-alfresco-viewer {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+
+    .mat-toolbar {
+        color: var(--theme-text-color);
+
+        .adf-toolbar-title {
+            width: auto;
+        }
+    }
+
+    &-main {
+        width: 0;
+    }
+
+    &__mimeicon {
+        vertical-align: middle;
+        height: 18px;
+        width: 18px;
+    }
+
+    &-toolbar {
+        .mat-toolbar {
+            background-color: var(--theme-card-bg-bold-color);
+        }
+    }
+
+    &__file-title {
+        text-align: center;
+    }
+
+    &__display-name {
+        font-size: var(--theme-subheading-2-font-size);
+        opacity: 0.87;
+        line-height: 1.5;
+        letter-spacing: -0.4px;
+        font-weight: normal;
+        font-style: normal;
+        font-stretch: normal;
+        max-width: 400px;
+        text-overflow: ellipsis;
+        overflow: hidden;
+        display: inline-block;
+        vertical-align: middle;
+        color: var(--theme-text-fg-color);
+    }
+
+    &-inline-container {
+        @extend .adf-full-screen;
+    }
+
+    &-custom-content {
+        width: 100vw;
+    }
+
+    &__sidebar {
+        width: 350px;
+        display: block;
+        padding: 0;
+        background-color: var(--theme-background-color);
+        box-shadow: 0 2px 4px 0 var(--theme-text-fg-shadow-color);
+        overflow: auto;
+
+        &__right {
+            border-left: 1px solid var(--theme-border-color);
+        }
+
+        &__left {
+            border-right: 1px solid var(--theme-border-color);
+        }
+    }
+
+}
