+++ lib/core/src/lib/viewer/components/img-viewer.component.html
@@ -13,7 +13,7 @@
             <mat-icon>zoom_out</mat-icon>
         </button>
 
-        <div class="adf-viewer__toolbar-page-scale" data-automation-id="adf-page-scale">
+        <div class="adf-image-viewer__toolbar-page-scale" data-automation-id="adf-page-scale">
             {{ currentScaleText }}
         </div>
