+++ lib/core/src/lib/viewer/components/img-viewer.component.scss
@@ -48,5 +48,20 @@
             display: inline-block;
             margin-left: 10px;
         }
+
+    }
+
+    &__toolbar {
+        &-page-scale {
+            cursor: default;
+            width: 79px;
+            height: 24px;
+            font-size: var(--theme-body-1-font-size);
+            border: 1px solid var(--theme-border-color);
+            text-align: center;
+            line-height: 24px;
+            margin-left: 4px;
+            margin-right: 4px;
+        }
     }
 }
