+++ lib/core/src/lib/viewer/components/media-player.component.html
@@ -1,4 +1,4 @@
 <video controls [ngClass]="{'adf-audio-file': mimeType && mimeType.startsWith('audio')}">
-    <source [src]="urlFile" [type]="mimeType" (error)="onMediaPlayerError()"/>
+    <source [src]="urlFile" [type]="mimeType" (error)="onMediaPlayerError($event)"/>
     <track *ngFor="let track of tracks" [kind]="track.kind" [label]="track.label" [srclang]="track.srclang" [src]="track.src"/>
 </video>
