+++ lib/core/src/lib/viewer/components/pdf-viewer-host.component.scss
@@ -593,8 +593,6 @@
         }
     }
 
-
-
     .adf-hidden, [hidden] {
         display: none !important;
     }
