+++ lib/core/src/lib/viewer/components/pdf-viewer.component.html
@@ -83,7 +83,7 @@
             <span>{{ 'ADF_VIEWER.PAGE_LABEL.OF' | translate }} {{ totalPages }}</span>
         </div>
 
-        <div class="adf-viewer__toolbar-page-scale" data-automation-id="adf-page-scale">
+        <div class="adf-pdf-viewer__toolbar-page-scale" data-automation-id="adf-page-scale">
             {{ currentScaleText }}
         </div>
