+++ lib/core/src/lib/viewer/components/pdf-viewer.component.ts
@@ -80,9 +80,6 @@ export class PdfViewerComponent implements OnChanges, OnDestroy {
     @Output()
     error = new EventEmitter<any>();
 
-    @Output()
-    close = new EventEmitter<any>();
-
     page: number;
     displayPage: number;
     totalPages: number;
@@ -516,8 +513,6 @@ export class PdfViewerComponent implements OnChanges, OnDestroy {
             .afterClosed().subscribe((password) => {
             if (password) {
                 callback(password);
-            } else {
-                this.close.emit();
             }
         });
     }
