+++ lib/core/src/lib/viewer/components/viewer-render.component.scss
@@ -1,0 +1,72 @@
+/* stylelint-disable scss/at-extend-no-missing-placeholder */
+.adf-full-screen {
+    width: 100%;
+    height: 100%;
+    background-color: var(--theme-card-bg-color);
+}
+
+.adf-viewer-render {
+
+    &-main {
+        width: 0;
+    }
+
+    &-content-container {
+        display: flex;
+        justify-content: center;
+    }
+
+    .adf-viewer-render-layout-content {
+        @extend .adf-full-screen;
+
+        position: relative;
+        overflow-y: hidden;
+        overflow-x: hidden;
+        z-index: 1;
+        background-color: var(--theme-background-color);
+
+        display: flex;
+        flex-direction: row;
+        flex-wrap: wrap;
+        flex: 1;
+
+        & > div {
+            display: flex;
+            flex-flow: row wrap;
+            margin: 0 auto;
+            align-items: stretch;
+            height: 93vh;
+            width: 100%;
+        }
+    }
+
+    &-overlay-container {
+        .adf-viewer-render-content {
+            position: fixed;
+            top: 0;
+            left: 0;
+            z-index: 1000;
+        }
+    }
+
+    &__loading-screen {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        flex-direction: column;
+        height: 85vh;
+
+        .mat-spinner {
+            margin: 0 auto;
+        }
+    }
+
+    &-custom-content {
+        width: 100vw;
+    }
+
+    &-unknown-content {
+        align-items: center;
+        display: flex;
+    }
+}
