--- lib/core/src/lib/viewer/components/viewer.component.html
@@ -1,142 +1,0 @@
-<div *ngIf="showViewer"
-     class="adf-viewer-container"
-     [class.adf-viewer-overlay-container]="overlayMode"
-     [class.adf-viewer-inline-container]="!overlayMode">
-
-    <div class="adf-viewer-content"
-         fxLayout="column"
-         [cdkTrapFocus]="overlayMode"
-         cdkTrapFocusAutoCapture>
-        <ng-content select="adf-viewer-toolbar"></ng-content>
-
-        <div fxLayout="row"
-             fxFlex="1 1 auto">
-            <ng-container *ngIf="allowRightSidebar && showRightSidebar">
-                <div class="adf-viewer__sidebar"
-                     [ngClass]="'adf-viewer__sidebar__right'"
-                     fxFlexOrder="4"
-                     id="adf-right-sidebar">
-<!--                    <ng-container *ngIf="sidebarRightTemplate">-->
-<!--                        <ng-container *ngTemplateOutlet="sidebarRightTemplate;context:sidebarRightTemplateContext">-->
-<!--                        </ng-container>-->
-<!--                    </ng-container>-->
-                    <ng-content *ngIf="!sidebarRightTemplate"
-                                select="adf-viewer-sidebar"></ng-content>
-                </div>
-            </ng-container>
-
-            <ng-container *ngIf="allowLeftSidebar && showLeftSidebar">
-                <div class="adf-viewer__sidebar"
-                     [ngClass]="'adf-viewer__sidebar__left'"
-                     fxFlexOrder="1"
-                     id="adf-left-sidebar">
-<!--                    <ng-container *ngIf="sidebarLeftTemplate">-->
-<!--                        <ng-container *ngTemplateOutlet="sidebarLeftTemplate;context:sidebarLeftTemplateContext">-->
-<!--                        </ng-container>-->
-<!--                    </ng-container>-->
-                    <ng-content *ngIf="!sidebarLeftTemplate"
-                                select="adf-viewer-sidebar"></ng-content>
-                </div>
-            </ng-container>
-
-            <div *ngIf="isLoading"
-                 class="adf-viewer-main"
-                 fxFlexOrder="1"
-                 fxFlex="1 1 auto">
-                <div class="adf-viewer-layout-content adf-viewer__fullscreen-container">
-                    <div class="adf-viewer-content-container">
-                        <ng-container *ngIf="isLoading">
-                            <div class="adf-viewer__loading-screen"
-                                 fxFlex="1 1 auto">
-                                <h2>{{ 'ADF_VIEWER.LOADING' | translate }}</h2>
-                                <div>
-                                    <mat-spinner></mat-spinner>
-                                </div>
-                            </div>
-                        </ng-container>
-
-                    </div>
-                </div>
-            </div>
-
-            <div *ngIf="!isLoading"
-                 class="adf-viewer-main"
-                 fxFlexOrder="1"
-                 fxFlex="1 1 auto">
-                <div class="adf-viewer-layout-content adf-viewer__fullscreen-container">
-                    <div class="adf-viewer-content-container" [ngSwitch]="viewerType">
-                        <ng-container *ngSwitchCase="'external'">
-<!--                            <adf-preview-extension-->
-<!--                                *ngIf="!!externalViewer"-->
-<!--                                [id]="externalViewer.component"-->
-<!--                                [node]="nodeEntry?.entry"-->
-<!--                                [url]="urlFileContent"-->
-<!--                                [extension]="externalViewer.fileExtension"-->
-<!--                                [attr.data-automation-id]="externalViewer.component">-->
-<!--                            </adf-preview-extension>-->
-                        </ng-container>
-
-                        <ng-container *ngSwitchCase="'pdf'">
-                            <adf-pdf-viewer (close)="onBackButtonClick()"
-                                            [thumbnailsTemplate]="thumbnailsTemplate"
-                                            [allowThumbnails]="allowThumbnails"
-                                            [blobFile]="blobFile"
-                                            [urlFile]="urlFileContent"
-                                            [fileName]="fileName"
-                                            [cacheType]="cacheTypeForContent"
-                                            (error)="onUnsupportedFile()"></adf-pdf-viewer>
-                        </ng-container>
-
-                        <ng-container *ngSwitchCase="'image'">
-                            <adf-img-viewer [urlFile]="urlFileContent"
-                                            [fileName]="fileName"
-                                            [blobFile]="blobFile"
-                                            (error)="onUnsupportedFile()"
-                                            (submit)="onSubmitFile($event)"
-                            ></adf-img-viewer>
-                        </ng-container>
-
-                        <ng-container *ngSwitchCase="'media'">
-                            <adf-media-player id="adf-mdedia-player"
-                                              [urlFile]="urlFileContent"
-                                              [mimeType]="mimeType"
-                                              [blobFile]="blobFile"
-                                              [fileName]="fileName"
-                                              (error)="onUnsupportedFile()"></adf-media-player>
-                        </ng-container>
-
-                        <ng-container *ngSwitchCase="'text'">
-                            <adf-txt-viewer [urlFile]="urlFileContent"
-                                            [blobFile]="blobFile"></adf-txt-viewer>
-                        </ng-container>
-
-                        <ng-container *ngSwitchCase="'custom'">
-                            <ng-container *ngFor="let ext of viewerExtensions">
-<!--                                <adf-preview-extension *ngIf="checkExtensions(ext.fileExtension)"-->
-<!--                                                       [id]="ext.component"-->
-<!--                                                       [node]="nodeEntry.entry"-->
-<!--                                                       [url]="urlFileContent"-->
-<!--                                                       [extension]="extension"-->
-<!--                                                       [attr.data-automation-id]="ext.component">-->
-<!--                                </adf-preview-extension>-->
-                            </ng-container>
-
-                            <span class="adf-viewer-custom-content"
-                                  *ngFor="let extensionTemplate of extensionTemplates">
-                                <ng-template *ngIf="extensionTemplate.isVisible"
-                                             [ngTemplateOutlet]="extensionTemplate.template"
-                                             [ngTemplateOutletContext]="{ urlFileContent: urlFileContent, extension:extension }">
-                                </ng-template>
-                            </span>
-                        </ng-container>
-
-                        <ng-container *ngSwitchDefault>
-                            <adf-viewer-unknown-format></adf-viewer-unknown-format>
-                        </ng-container>
-                    </div>
-                </div>
-            </div>
-
-        </div>
-    </div>
-</div>
