--- lib/core/src/lib/viewer/components/viewer.component.scss
@@ -1,205 +1,0 @@
-/* stylelint-disable scss/at-extend-no-missing-placeholder */
-.adf-full-screen {
-    width: 100%;
-    height: 100%;
-    background-color: var(--theme-card-bg-color);
-}
-
-.adf-viewer {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-
-    .mat-toolbar {
-        color: var(--theme-text-color);
-
-        .adf-toolbar-title {
-            width: auto;
-        }
-    }
-
-    &-main {
-        width: 0;
-    }
-
-    &__mimeicon {
-        vertical-align: middle;
-        height: 18px;
-        width: 18px;
-    }
-
-    &-toolbar {
-        .mat-toolbar {
-            background-color: var(--theme-card-bg-bold-color);
-        }
-    }
-
-    &__file-title {
-        text-align: center;
-    }
-
-    &__display-name {
-        font-size: var(--theme-subheading-2-font-size);
-        opacity: 0.87;
-        line-height: 1.5;
-        letter-spacing: -0.4px;
-        font-weight: normal;
-        font-style: normal;
-        font-stretch: normal;
-        max-width: 400px;
-        text-overflow: ellipsis;
-        overflow: hidden;
-        display: inline-block;
-        vertical-align: middle;
-        color: var(--theme-text-fg-color);
-    }
-
-    &-container {
-        .adf-viewer-layout-content {
-            @extend .adf-full-screen;
-
-            position: relative;
-            overflow-y: hidden;
-            overflow-x: hidden;
-            z-index: 1;
-            background-color: var(--theme-background-color);
-            display: flex;
-            flex-direction: row;
-            /* stylelint-disable-next-line declaration-block-no-redundant-longhand-properties */
-            flex-wrap: wrap;
-            flex: 1;
-
-            & > div {
-                display: flex;
-                flex-flow: row wrap;
-                margin: 0 auto;
-                align-items: stretch;
-                height: 100%;
-            }
-        }
-
-        .adf-viewer-layout {
-            @extend .adf-full-screen;
-
-            display: flex;
-            flex-direction: row;
-            overflow-y: auto;
-            overflow-x: hidden;
-            position: relative;
-        }
-
-        .adf-viewer-content {
-            @extend .adf-full-screen;
-
-            flex: 1;
-
-            & > div {
-                height: 0; // Firefox
-            }
-        }
-    }
-
-    &-overlay-container {
-        .adf-viewer-content {
-            position: fixed;
-            top: 0;
-            left: 0;
-            z-index: 1000;
-        }
-    }
-
-    &-inline-container {
-        @extend .adf-full-screen;
-    }
-
-    &-content-container {
-        display: flex;
-        justify-content: center;
-    }
-
-    &-custom-content {
-        width: 100vw;
-    }
-
-    &-unknown-content {
-        align-items: center;
-        display: flex;
-    }
-
-    &__loading-screen {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        flex-direction: column;
-        height: 85vh;
-
-        .mat-spinner {
-            margin: 0 auto;
-        }
-    }
-
-    &__sidebar {
-        width: 350px;
-        display: block;
-        padding: 0;
-        background-color: var(--theme-background-color);
-        box-shadow: 0 2px 4px 0 var(--theme-text-fg-shadow-color);
-        overflow: auto;
-
-        &__right {
-            border-left: 1px solid var(--theme-border-color);
-        }
-
-        &__left {
-            border-right: 1px solid var(--theme-border-color);
-        }
-    }
-
-    &__thumbnails {
-        width: 180px;
-        display: flex;
-        flex-direction: column;
-        padding: 0;
-        background: #e6e6e6;
-
-        .adf-info-drawer-layout {
-            display: flex;
-            flex-direction: column;
-            flex: 1;
-            background: #e6e6e6;
-        }
-
-        .adf-info-drawer-layout-header {
-            margin-bottom: 0;
-        }
-
-        .adf-info-drawer-layout-content {
-            padding: 0;
-            height: 100%;
-            overflow: hidden;
-        }
-
-        .adf-info-drawer-content {
-            height: 100%;
-        }
-
-        .adf-info-drawer-layout-content > *:last-child {
-            height: 100%;
-            overflow: hidden;
-        }
-    }
-
-    &__toolbar {
-        &-page-scale {
-            cursor: default;
-            width: 79px;
-            height: 24px;
-            font-size: var(--theme-body-1-font-size);
-            border: 1px solid var(--theme-border-color);
-            text-align: center;
-            line-height: 24px;
-            margin-left: 4px;
-            margin-right: 4px;
-        }
-    }
-}
